+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Shortcut mode trigger changed from Space (when empty) to Ctrl+Space (works anytime)
+
 ## [0.0.0] - 2026-01-23
 
 ### Added